requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pip-tools==7.4.1
pluggy==1.5.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
